% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGroup.R
\name{addGroup}
\alias{addGroup}
\title{Select background cells (or samples) and assign groups}
\usage{
addGroup(diffMIN_object, group)
}
\arguments{
\item{diffMIN_object}{A \code{diffMIN_object} S4 object containing raw data and metadata.}

\item{group}{A character vector specifying the column names of background cells (or samples).}
}
\value{
The updated \code{diffMIN_object} S4 object with an additional \code{group} column in \code{meta.data}, indicating whether each cell (or sample) is a background ('bg') or foreground ('fg').
}
\description{
This function classifies cells (or samples) into background ('bg') and foreground ('fg') groups based on the provided column names.
}
\examples{
library(diffMIN)
library(stringr)
data(exp)
obj <- create_diffMIN_object(rawdata = exp)
group <- colnames(exp)[str_detect(colnames(exp), "Sham")]
obj <- addGroup(obj, group)
}
