% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diffMIN_object.R
\name{$.diffMIN_object}
\alias{$.diffMIN_object}
\title{Extract method for $ operator for diffMIN_object}
\usage{
\method{$}{diffMIN_object}(x, name)
}
\arguments{
\item{x}{Object of class diffMIN_object}

\item{name}{Character string specifying the column name in meta.data}
}
\value{
The column of meta.data corresponding to 'name'
}
\description{
Extract method for $ operator for diffMIN_object
}
