% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Data in a \code{diffMIN} Object}
\usage{
cluster(
  diffMIN_object,
  reduction.use = c("umap", "tsne", "pca"),
  centers,
  method = c("hc", "kmean")
)
}
\arguments{
\item{diffMIN_object}{An object of class \code{diffMIN} containing dimensionality-reduced data.}

\item{reduction.use}{A character string specifying which reduction method to use. Options are \code{"umap"}, \code{"tsne"}, or \code{"pca"}.}

\item{centers}{An integer specifying the number of clusters for the clustering algorithm.}

\item{method}{A character string specifying the clustering method. Options are \code{"hc"} for hierarchical clustering and \code{"kmean"} for k-means clustering.}
}
\value{
The modified \code{diffMIN} object with an additional column \code{cluster} in the \code{meta.data} slot, which contains the cluster assignments.
}
\description{
This function performs clustering on the dimensionality-reduced data stored in a \code{diffMIN} object.
}
\examples{
\dontrun{
library(diffMIN)
data(exp)
obj <- create_diffMIN_object(rawdata = exp)
obj <- cluster(diffMIN_object = obj, reduction.use = "umap", centers = 3, method = "kmean")
}
}
