% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NMI_network.R
\name{compute_NMI_network}
\alias{compute_NMI_network}
\title{Compute Background or Foreground NMI Network}
\usage{
compute_NMI_network(
  diffMIN_obj = obj,
  type = NULL,
  cores = "Max",
  batch_size = 30,
  p.value.cutoff = 0.05,
  permutation.test.times = 1000,
  file.dir = "./",
  filename = "diffMIN_obj",
  whether_to_save = TRUE,
  use_adj.p = FALSE
)
}
\arguments{
\item{diffMIN_obj}{An object of class \code{diffMIN} containing the necessary data for network computation.}

\item{type}{A character string indicating the type of network to compute. Use \code{"bg"} for background network or \code{"fg"} for foreground network.}

\item{cores}{The number of CPU cores to use for parallel processing. Use \code{"Max"} to utilize all available cores.}

\item{batch_size}{An integer specifying the number of samples processed per batch.}

\item{p.value.cutoff}{A numeric threshold for significant edges. Edges with p-values below this cutoff are considered significant.}

\item{permutation.test.times}{The number of permutations to perform for the permutation test.}

\item{file.dir}{A character string specifying the directory to save the RData file.}

\item{filename}{A character string specifying the name of the RData file.}

\item{whether_to_save}{A logical value indicating whether to save the \code{diffMIN_obj} to file.}

\item{use_adj.p}{Whether use BH method to adjust p value.}
}
\value{
An updated \code{diffMIN} object with computed networks added to the \code{net.EM} slot.
}
\description{
This function computes the background or foreground network using Normalized Mutual Information (NMI) on data stored in a \code{diffMIN} object. The function supports parallel processing and batch computations.
}
\examples{
\dontrun{
library(diffMIN)
data(exp)
obj <- create_diffMIN_object(rawdata = exp)
obj <- compute_NMI_network(
  diffMIN_obj = obj,
  type = "bg",
  cores = "Max",
  batch_size = 30,
  p.value.cutoff = 0.05,
  permutation.test.times = 1000,
  file.dir = "./",
  filename = "diffMIN_obj",
  whether_to_save = TRUE,
  use_adj.p = FALSE
)
}
}
