% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diffMIN_object.R
\name{create_diffMIN_object}
\alias{create_diffMIN_object}
\title{Create a diffMIN_object S4 Object}
\usage{
create_diffMIN_object(
  rawdata,
  dat = data.frame(),
  net.DM = data.frame(),
  net.EM = list(bg = list(), fg = list()),
  reduction = list(),
  meta.data = data.frame(),
  IN = list(bg.all.expr = list(), fg.all.expr = list(), bg.ran.net.up = list(),
    bg.ran.net.down = list(), fg.ran.net.up = list(), fg.ran.net.down = list())
)
}
\arguments{
\item{rawdata}{A matrix containing the raw expression data.}

\item{dat}{A matrix containing the discretized expression data.}

\item{net.DM}{A data frame representing the diffMIN degree matrix.}

\item{net.EM}{A list containing the diffMIN edge matrices. Should include elements \code{bg} and \code{fg}, each being a list.}

\item{reduction}{A list containing dimensionality reduction results (e.g., UMAP, PCA, t-SNE).}

\item{meta.data}{A data frame containing metadata associated with the samples.}

\item{IN}{A list containing intermediate network data used for final network calculations.}
}
\value{
An object of class \code{diffMIN_object}.
}
\description{
This function creates an S4 object of class \code{diffMIN_object}.
}
\examples{
\dontrun{
data(exp)
exp <- pre_process(exp)
obj <- create_diffMIN_object(rawdata = exp)
print(obj)
}
}
