% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffMIN_class.R
\docType{class}
\name{diffMIN_object}
\alias{diffMIN_object}
\alias{diffMIN_object-class}
\title{diffMIN_object Class Definition}
\arguments{
\item{net.DM}{Data frame: The diffMIN degree matrix.}

\item{net.EM}{list: The diffMIN edge matrix.}
}
\description{
diffMIN_object Class Definition
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{The raw expression matrix.}

\item{\code{dat}}{The discretized expression matrix.}

\item{\code{reduction}}{Reduction data containing UMAP, PCA, t-SNE, etc.}

\item{\code{meta.data}}{Metadata associated with the data.}

\item{\code{IN}}{Intermediate network used for calculating the final network}
}}

