% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_EM.R
\name{discretize_EM}
\alias{discretize_EM}
\title{Discretize expression matrix}
\usage{
discretize_EM(diffMIN_object, ...)
}
\arguments{
\item{diffMIN_object}{A diffMIN_object S4 Object}
}
\value{
A diffMIN_object S4 Object
}
\description{
Discretize expression matrix
}
\examples{
\dontrun{
library(diffMIN)
data(exp)
exp <- pre_process(exp)
obj <- create_diffMIN_Object(rawdata = exp)
obj <- discretize_EM(obj)
}
}
