% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\docType{data}
\name{exp}
\alias{exp}
\title{Single Cell Sequencing of Cardiomyocytes in Mouse TAC Surgical Model}
\format{
\subsection{\code{exp}}{

A data frame with 22,134 rows and 482 columns:
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE95140}
}
\usage{
exp
}
\description{
A subset of data from NCBI GEO dataset GSE95140 which bulk columns are removed
Report ...
}
\keyword{datasets}
