% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMI.R
\name{NMI}
\alias{NMI}
\title{Calculate Normalized Mutual Information (NMI)}
\usage{
NMI(dis, ...)
}
\arguments{
\item{dis}{A data frame or matrix where each column represents a discrete variable.}
}
\value{
A named vector containing the NMI values for each pair of variables.
}
\description{
This function computes the normalized mutual information (NMI) between each pair of columns in a data frame or matrix.
}
\examples{
# Example usage
dis <- data.frame(A = sample(1:3, 10, replace = TRUE), B = sample(1:3, 10, replace = TRUE))
NMI(dis)
}
