% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\docType{data}
\name{obj}
\alias{obj}
\title{Single Cell Sequencing of Cardiomyocytes in Mouse TAC Surgical Model}
\format{
\subsection{\code{obj}}{

A diffMIN_object with 400 rows and 396 columns:
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE95140}
}
\usage{
obj
}
\description{
A subset of data from NCBI GEO dataset GSE95140 which bulk columns are removed
Report ...
}
\keyword{datasets}
