% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDot.R
\name{plotDot}
\alias{plotDot}
\title{Plot Dot Plot for diffMIN_object}
\usage{
plotDot(
  diffMIN_object,
  reduction = c("umap", "tsne", "pca"),
  group.by = "group"
)
}
\arguments{
\item{diffMIN_object}{An object of class \code{diffMIN_object}.}

\item{reduction}{Character string specifying the reduction method to use. Options are "umap", "tsne", or "pca".}

\item{group.by}{Character string specifying the column name in \code{meta.data} to use for grouping.}
}
\value{
A ggplot object displaying the dot plot.
}
\description{
This function creates a dot plot of the dimensionality-reduced data in a \code{diffMIN_object} and colors the points based on a specified grouping from the metadata.
}
\examples{
\dontrun{
library(diffMIN)
data(obj)
obj <- transform_EM_to_DM(obj)
obj <- RunReduc(diffMIN_object = obj,reduction = c("umap","tsne","pca"),seed = 1)
plotDot(obj, reduction = "tsne", group.by = "group")
}
}
