% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHM.R
\name{plotHM}
\alias{plotHM}
\title{Plot Heatmap for diffMIN_object}
\usage{
plotHM(diffMIN_object, group.by = "group", type = "degree", scale = TRUE)
}
\arguments{
\item{diffMIN_object}{An object of class \code{diffMIN_object}.}

\item{group.by}{Character string specifying the column in \code{meta.data} to use for grouping.}

\item{type}{Character string specifying the type of data to plot. Options are "degree", "exp", or "edge".}

\item{scale}{Logical indicating whether to scale the data.}
}
\value{
A heatmap plot created using \code{ComplexHeatmap}.
}
\description{
This function generates a heatmap based on the specified type of data from a \code{diffMIN_object}.
}
\examples{
# Example usage
library(ComplexHeatmap)
data(exp)
exp <- pre_process(exp)
obj <- create_diffMIN_object(exp)
plotHM(obj, group.by = "group", type = "exp")
}
