% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\title{Pre-process Expression Matrix}
\usage{
pre_process(
  exp,
  minExpr_sample = c(0.1, 5000),
  minExpr_gene = c(0, 20),
  var.genes = 400
)
}
\arguments{
\item{exp}{A data.frame or matrix. The expression data where rows are genes and columns are samples.}

\item{minExpr_sample}{A numeric vector of length 2. The first value is the minimum expression level required per sample,
and the second value is the minimum number of genes that must exceed this expression level for a sample
to be retained. Default is c(0.1, 5000).}

\item{minExpr_gene}{A numeric vector of length 2. The first value is the minimum expression level required per gene,
and the second value is the minimum number of samples that must exceed this expression level for a gene
to be retained. Default is c(0, 20).}

\item{var.genes}{Integer. The number of highly variable genes to select. Default is 400.}
}
\value{
A data frame with pre-processed expression data.
}
\description{
This function pre-processes expression data by removing low expression cells (or samples),
removing low expression genes, and selecting highly variable genes. It also provides a warning
if the number of cells (or samples) exceeds 1000, recommending a reduction for better performance.
}
\examples{
data(exp)
dim(exp)
result <- pre_process(exp)
dim(result)
}
