% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_network.R
\name{prepare_network}
\alias{prepare_network}
\title{Prepare Data for Computing Network}
\usage{
prepare_network(diffMIN_object, nRanNet = 100, seed = 123)
}
\arguments{
\item{diffMIN_object}{An object of class \code{diffMIN_object}.}

\item{nRanNet}{Integer: The number of random networks to generate.}

\item{seed}{Integer: Random seed for reproducibility.}
}
\value{
An updated \code{diffMIN_object} with prepared networks.
}
\description{
This function prepares random background and foreground networks for a \code{diffMIN_object}.
}
\examples{
# Example usage
data(exp)
exp <- pre_process(exp)
obj <- create_diffMIN_object(exp)
obj <- prepare_network(obj, nRanNet = 100, seed = 123)
}
