% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diffMIN_object.R
\name{print.diffMIN_object}
\alias{print.diffMIN_object}
\title{Print method for diffMIN_object}
\usage{
\method{print}{diffMIN_object}(x, ...)
}
\arguments{
\item{x}{Object of class diffMIN_object}

\item{...}{Additional arguments (currently ignored)}
}
\description{
Print method for diffMIN_object
}
