% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_EM_to_DM.R
\name{transform_EM_to_DM}
\alias{transform_EM_to_DM}
\title{Transform Edge Matrix to Degree Matrix}
\usage{
transform_EM_to_DM(diffMIN_object)
}
\arguments{
\item{diffMIN_object}{An object of class diffMIN_object.}
}
\value{
An object of class diffMIN_object with updated degree matrices.
}
\description{
Transform Edge Matrix to Degree Matrix
}
