% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDegree.R
\name{plotDegree}
\alias{plotDegree}
\title{Plot Dotplot for Gene Expression}
\usage{
plotDegree(diffMIN_object, feature, reduction = "tsne")
}
\arguments{
\item{diffMIN_object}{An object of class \code{diffMIN_object}.}

\item{feature}{Character string specifying the feature (degree) to plot.}

\item{reduction}{Character string specifying the reduction result to use. Options are "umap", "tsne", or "pca".}
}
\value{
A ggplot object showing the dotplot of gene expression.
}
\description{
Plot Dotplot for Gene Expression
}
