% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runReduc.R
\name{RunReduc}
\alias{RunReduc}
\title{Run reduction for diffMIN_object}
\usage{
RunReduc(
  diffMIN_object,
  reduction = c("umap", "tsne", "pca"),
  seed = 1,
  perplexity = 4,
  ...
)
}
\arguments{
\item{diffMIN_object}{An object of class diffMIN_object.}

\item{reduction}{Methods to perform}

\item{seed}{A random seed for performing reduction}

\item{perplexity}{A parameter for tsne}

\item{...}{Some parameters for umap}

\item{slot}{Edge matrix or degree matrix, degree matrix is defaulted}
}
\value{
An updated \code{diffMIN} object with reduction value added to the \code{reduction} slot.
}
\description{
Run reduction for diffMIN_object
}
\examples{
\dontrun{
library(diffMIN)
data(obj)
obj <- transform_EM_to_DM(obj)
obj <- RunReduc(diffMIN_object = obj,reduction = c("umap","tsne","pca"),seed = 1)
}
}
